+++ tests/behat/features/search-for-a-page.feature
@@ -1,3 +1,4 @@
+@retry
 Feature: Search for a page
   As an author
   I want to search for a page in the CMS
